requirements.txt
tkinter
customtkinter
geopy
bs4
requests
pillow
json
